==> scripts/01_ProcessCounts.Rmd <==
---
title: "Process Counts"
output: html_notebook
---

Creating B. rapa networks from eQTL data to enhance Rob Baker et al's FVT analysis for stem elongation.

### set up
```{r set_up}
library(edgeR)
#library(boxr)
library(magrittr)
#box_auth()
#box_setwd(24503987860)
#box_ls() %>% as.data.frame()
library(rio)
```

## Get data
```{r}
#counts <- box_read_tsv(162218089685)
counts <- read.csv("../input/RIL_v1.5_mapping_UN.csv.gz")
row.names(counts) <- counts$gene
counts <- counts [-1,-1] #remove first row (umapped reads) and first column (gene IDs)
counts[is.na(counts)] <- 0
head(counts[,1:10])
summary(counts[,1:10])
```


### set up descriptive data frame
```{r}
pdata <- colnames(counts) %>% data.frame(file=.,
                    RIL=regmatches(.,regexpr("RIL_[0-9]*",.)),
                    trt=regmatches(.,regexpr("CR|UN",.)),
                    rep=regmatches(.,regexpr("Rep[0-9]*",.)),
                    stringsAsFactors = FALSE
)
pdata$group <- paste(pdata$RIL,pdata$trt,sep="_")
head(pdata)
tail(pdata)
sort(unique(pdata$RIL))
```


## only UN reads used for paper

### set up descriptive data frame

```{r}
counts.UN <- counts[,pdata$trt=="UN",]

pdata.UN <- pdata %>% subset(trt=="UN")
```

```{r}
counts.dge.UN <- DGEList(counts=counts.UN,group = pdata.UN$group)
counts.dge.UN$samples
```

Filter UN
```{r}
nrow(counts.dge.UN)
counts.dge.UN <- counts.dge.UN[rowSums(cpm(counts.dge.UN)>2) >= 44,,keep.lib.sizes=FALSE]
counts.dge.UN$samples
nrow(counts.dge.UN)
```
Kept 28,668 genes of 43,150

```{r}
counts.dge.UN <- calcNormFactors(counts.dge.UN)
counts.dge.UN$samples
counts.log.cpm.UN <- cpm(counts.dge.UN,log=TRUE)
#box_write(counts.log.cpm.UN,"RIL_UN_TMM_CPM.csv.gz")
export(counts.log.cpm.UN,"~/Box Sync/BrapaNetworks/RIL_UN_TMM_CPM.csv.gz")
```

estimate dispersions
```{r}
design.UN <- model.matrix(~pdata.UN$RIL)

# #estimate Dispersion hangs...
#counts.dge.CR <- estimateDisp(counts.dge.CR,design = design.CR)

#counts.dge.UN <- estimateDisp(counts.dge.UN,design = design.UN)
```

```{r}
#box_auth()
#box_save(counts,counts.dge,file_name = "NormalizedCounts.RData")
#save(counts.dge.CR,counts.dge.UN,file="~/Box Sync/BrapaNetworks/Normalized_CR_UN_Counts.Rdata")
```

```{r}
counts.voom.UN <- voom(counts.dge.UN,design = design.UN)
```

```{r}
voom.df.UN <- as.data.frame(counts.voom.UN)
write.csv(voom.df.UN,file="~/Box Sync/BrapaNetworks/voom.UN.csv",row.names = TRUE)
write.csv(voom.df.UN,file="../output/voom.UN.csv",row.names = TRUE)
#box_write(voom.df,filename = "voom.csv.gz")
```

Fitted values additiive
```{r}
voom.fit.UN <- lmFit(counts.voom.UN,design=design.UN)
voom.fit.UN <- eBayes(voom.fit.UN)
voom.topTable.UN <- topTableF(voom.fit.UN,number=Inf)
save(counts.voom.UN, voom.fit.UN, voom.topTable.UN, file="~/Box Sync/BrapaNetworks/voom_UN.Rdata")
save(counts.voom.UN, voom.fit.UN, voom.topTable.UN, file="../output/voom_UN.Rdata")
```
