==> scripts/02_MR Clustering UN.Rmd <==
---
title: "UN Mutual Rank Clustering"
output: html_notebook
---

Notebook for generating mutual rank (MR) clusters from the UN data.

```{r}
library(boxr)
library(magrittr)
library(knitr)
library(limma)
library(snowfall)
library(igraph)
library(plyr)
library(tidyverse)
#box_auth()
#box_setwd(24503987860)
#box_ls()
```

### First attempt, use RIL means
Get data
```{r}
#box_load(162362592785)
load("../output/Voom_UN.Rdata")
annotation  <- read.csv("../input/Brapa_V1.5_annotated.csv")
```


```{r}
RIL.coefs.UN <- grep("RIL",colnames(coef(voom.fit.UN)))
RIL.pvals.UN <- topTable(voom.fit.UN,coef = RIL.coefs.UN,number = Inf)
nrow(RIL.pvals.UN)
sum(RIL.pvals.UN$adj.P.Val<0.01)
sum(RIL.pvals.UN$adj.P.Val<10e-05) 
```

Start by taking the top 10000

```{r}
genes.of.interest.UN <- row.names(RIL.pvals.UN)[1:10000]
expr.data.UN <- voom.fit.UN$coefficients[genes.of.interest.UN,]
dim(expr.data.UN)
head(expr.data.UN[,1:6])
```


```{r}
expr.data.UN[,-1] <- expr.data.UN[,-1] + expr.data.UN[,1] # add the intercept to the coefficients
colnames(expr.data.UN) <- sub("pdata.UN$RIL","",colnames(expr.data.UN),fixed=TRUE)
colnames(expr.data.UN)[1] <- "RIL_1"
head(expr.data.UN[,1:6])
```

Get Rob's Data.  Remove STP trait and other not-needed columns
```{r}
blups2011 <- read.csv("../input/Brapa2011BayesHeight_blups.csv",row.names = 1)
blups2012 <- read.csv("../input/Brapa2012BayesHeight_blups.csv", row.names = 1)
blups2011$Line %<>% paste("RIL",.,sep="_")
blups2012$Line %<>% paste("RIL",.,sep="_")
blups2011 %<>% subset(select=!grepl("V1|individual|blk|trt|treat|STP",colnames(.))) %>%
  rename(UNiD=UNInflect_DD,UNd=UNduration) %>%
  select(starts_with("Line"),starts_with("UN"),-matches("Inflect.*_size"))
blups2012 %<>% subset(select=!grepl("V1|individual|blk|trt|treat|STP",colnames(.))) %>%
  rename(UNiD=UNInflection_DD,UNd=UNduration) %>%
    select(starts_with("Line"),starts_with("UN"),-matches("Inflect.*_size"))
head(blups2011)
head(blups2012)
```

Have decided to use 2011 and 2012 UN BLUPs.
```{r}
blups.UN <- merge(blups2011,blups2012,by="Line",suffixes = c("11","12"))
head(blups.UN)
```

Merge blups with expression
```{r}
expr.blup.UN <- merge(blups.UN,t(expr.data.UN),by.x="Line",by.y=0)
head(expr.blup.UN[,1:6])
```


Correlation and Mutual Rank
```{r}
sfInit(parallel = TRUE, cpus=4)

expr.blup.cor.UN <- cor(expr.blup.UN[,-1])

expr.blup.rank.UN.all <- sfApply(expr.blup.cor.UN,2,function(x) rank(-abs(x)))
expr.blup.rank.UN.all[1:10,1:10]
expr.blup.mr.UN.all <- sqrt(expr.blup.rank.UN.all*t(expr.blup.rank.UN.all))
expr.blup.mr.UN.all[1:10,1:10]

expr.blup.rank.UN.2011 <- expr.blup.UN %>% 
  select(ends_with("11"), starts_with("Bra"))  %>%
  cor() %>% 
  sfApply(2,function(x) rank(-abs(x)))
expr.blup.rank.UN.2011[1:10,1:10]
expr.blup.mr.UN.2011 <- sqrt(expr.blup.rank.UN.2011*t(expr.blup.rank.UN.2011))
expr.blup.mr.UN.2011[1:10,1:10]

expr.blup.rank.UN.2012 <- expr.blup.UN %>% 
  select(ends_with("12"), starts_with("Bra"))  %>%
  cor() %>% 
  sfApply(2,function(x) rank(-abs(x)))
expr.blup.rank.UN.2012[1:10,1:10]
expr.blup.mr.UN.2012 <- sqrt(expr.blup.rank.UN.2012*t(expr.blup.rank.UN.2012))
expr.blup.mr.UN.2012[1:10,1:10]

sfStop()
```


```{r}
save(blups.UN, expr.blup.UN, expr.blup.mr.UN.all, expr.blup.mr.UN.2012, expr.blup.mr.UN.2011,expr.data.UN, expr.blup.cor.UN,file = "../output/MR_UN.Rdata")
#save(blups.UN, expr.blup.UN, expr.blup.mr.UN.all, expr.blup.mr.UN.2012, expr.blup.mr.UN.2011,expr.data.UN, expr.blup.cor.UN,file = "~/Box Sync/BrapaNetworks/MR_UN.Rdata")
#file.copy(from = "~/Box Sync/BrapaNetworks/MR_UN.Rdata", to= "../output/MR_UN.Rdata",overwrite = TRUE)
```

Get graphs at different MR

## get data

```{r}
load("../output/MR_UN.Rdata")
annotation <- read.csv("../input/Brapa_V1.5_annotated.csv",row.names=1,as.is = TRUE)
```

## define some functions

```{r mr_subgraphs, echo=FALSE, cache.lazy=FALSE, warning=FALSE, eval=TRUE }
get.mr.subgraph <- function(mr.cutoff,mr.matrix,annotation=NULL,neighborhood,order=1) {
  #function to extract graph at a specificed correlation cutoff
  gene.mr.tmp <- mr.matrix
  gene.mr.tmp[abs(gene.mr.tmp) > mr.cutoff] <- 0
  gene.mr.tmp[is.na(gene.mr.tmp)] <- 0 #important! otherwise vertices with NA edges are connected
  
  gene.graph <- graph.adjacency(adjmatrix = gene.mr.tmp,
                                mode="undirected",
                                weighted="mr",
                                diag=FALSE)
  
  sub.graphs <- graph.neighborhood(gene.graph,order=order,nodes=neighborhood) #for each node of interest get all other nodes within order of 1
  
  #get combined list of vertices...
  
  sub.vertices <- unique(names(unlist(lapply(sub.graphs, V))))
  
  combined.sub.graph <- induced_subgraph(gene.graph,sub.vertices)
  
  V(combined.sub.graph)$color <- "lightblue"
  V(combined.sub.graph)[neighborhood]$color <- "red"
  if(!is.null(annotation)) V(combined.sub.graph)$gene <- 
    annotation$At_symbol[match(V(combined.sub.graph)$name,annotation$name)]
  list(cutoff=mr.cutoff,graph=combined.sub.graph,nodes=V(combined.sub.graph)$name)
}
```

## Get blup networks at different MR thresholds
Use different MR thresholds of 10 to 50 (absolute values) to build the network (genes with a correlation higher than the threshold were considered connected).  Then subset the network to take all genes "directly" connected to a FVT

```{r, echo=FALSE, eval=TRUE}
cutoffs <- c(15,25,35,50)

blup.mr.graphs.UN.all <- lapply(cutoffs, get.mr.subgraph, mr.matrix = expr.blup.mr.UN.all, annotation= annotation, neighborhood = colnames(blups.UN)[-1])
names(blup.mr.graphs.UN.all) <- sapply(blup.mr.graphs.UN.all, function(x) paste("blup.mr.graph.UN.all",x$cutoff,sep="."))
```

```{r, echo=FALSE, eval=TRUE}
cutoffs <- c(10,20,30,50)

blup.mr.graphs.UN.2011 <- lapply(cutoffs, get.mr.subgraph, mr.matrix = expr.blup.mr.UN.2011, annotation= annotation, neighborhood = colnames(blups.UN) %>% str_subset("11"))
names(blup.mr.graphs.UN.2011) <- sapply(blup.mr.graphs.UN.2011, function(x) paste("blup.mr.graph.UN.2011",x$cutoff,sep="."))
```

```{r, echo=FALSE, eval=TRUE}
cutoffs <- c(10,20,30,50)

blup.mr.graphs.UN.2012 <- lapply(cutoffs, get.mr.subgraph, mr.matrix = expr.blup.mr.UN.2012, annotation= annotation, neighborhood = colnames(blups.UN) %>% str_subset("12"))
names(blup.mr.graphs.UN.2012) <- sapply(blup.mr.graphs.UN.2012, function(x) paste("blup.mr.graph.UN.2012",x$cutoff,sep="."))
```

## get edge correlation info

```{r}
edge.col.sign <- function(graph,edge.colors) {#color edges by correlation sign
  E(graph) %>% 
    attr("vnames") %>% 
    tibble(edge=.) %>% 
    separate(edge,into=c("node1","node2"),remove = FALSE,sep="\\|") %>%
    rowwise() %>% 
    mutate(cor=expr.blup.cor.UN[node1,node2],
           col=ifelse(cor > 0, edge.colors[1], edge.colors[2])) %>%
    select(col) %>%
    unlist()
}
```


## plots

```{r plot.function}
plot.graph <- function (blup.graph) {
  if(blup.graph$cutoff <= cutoff) {
    E(blup.graph$graph)$width <- rank(E(blup.graph$graph)$mr)^(1/4)
    E(blup.graph$graph)$color <- edge.col.sign(blup.graph$graph, edge.colors=c("magenta","green"))
    plot(blup.graph$graph,
         layout = layout_with_kk, 
         vertex.label = ifelse(is.na(V(blup.graph$graph)$gene)|V(blup.graph$graph)$gene=="",
                               V(blup.graph$graph)$name,
                               paste(V(blup.graph$graph)$gene,V(blup.graph$graph)$name,sep="\n")),
         vertex.label.cex=1,
         main=paste("MR cutoff =",sub("blup.mr.graph.","",blup.graph$cutoff,fixed=T)))
    rm(blup.graph)
  }
}
```

All
```{r plot_all, echo=FALSE, cache=TRUE}
cutoff=200
pdf("../output/blup.mr.plots.UN.all.pdf",height=12,width=12)
result <- sapply(blup.mr.graphs.UN.all,plot.graph)
dev.off()
file.copy(from = "../output/blup.mr.plots.UN.all.pdf", to = "~/Box Sync/BrapaNetworks/blup.mr.plots.UN.all.pdf", overwrite = TRUE)
```

2011
```{r plot_2011, echo=FALSE, cache=TRUE}
cutoff=200
pdf("../output/blup.mr.plots.UN.2011.pdf",height=12,width=12)
result <- sapply(blup.mr.graphs.UN.2011,plot.graph)
dev.off()
file.copy(from = "../output/blup.mr.plots.UN.2011.pdf", to = "~/Box Sync/BrapaNetworks/blup.mr.plots.UN.2011.pdf", overwrite = TRUE)
```

2012
```{r plot_2012, echo=FALSE, cache=TRUE}
cutoff=200
pdf("../output/blup.mr.plots.UN.2012.pdf",height=12,width=12)
result <- sapply(blup.mr.graphs.UN.2012,plot.graph)
dev.off()
file.copy(from = "../output/blup.mr.plots.UN.2012.pdf", to = "~/Box Sync/BrapaNetworks/blup.mr.plots.UN.2012.pdf", overwrite = TRUE)
```

### make annotation tables for plots
```{r annoation.functions}
annotate.nodes <- function(graph.list) {
  tmp.results <- lapply(graph.list, function (blup.graph) {
    if(blup.graph$cutoff <= cutoff) {
      data.frame(
        MR_Cutoff=blup.graph$cutoff,
        trt="UN",
        annotation[annotation$name %in% blup.graph$nodes,]
      )
    }
  }
  )
  
  tmp.results <- ldply(tmp.results) 
  
  tmp.results[!duplicated(tmp.results$name),]
}

annotate.edges <- function(graph.list) {
  tmp.results <- map(graph.list, function(blup.graph) {
    if(blup.graph$cutoff <= cutoff) {
      E(blup.graph$graph) %>% 
        attr("vnames") %>% 
        tibble(edge=.) %>% 
        separate(edge,into=c("node1","node2"),remove = FALSE,sep="\\|") %>%
        group_by(edge) %>% 
        summarize(MR_Cutoff=blup.graph$cutoff, node1, node2, cor=expr.blup.cor.UN[node1,node2])
    }
  }
  )
  
  bind_rows(tmp.results, .id = "graph") %>% filter(!duplicated(edge))
}
```


Node annotation
```{r}
blup.mr.graphs.UN.all %>% annotate.nodes %>% 
  write_csv(path="../output/MR_UN_graphs_node_annotation_all.csv")

blup.mr.graphs.UN.2011 %>% annotate.nodes %>%
  write_csv(path="../output/MR_UN_graphs_node_annotation_2011.csv")

blup.mr.graphs.UN.2012 %>% annotate.nodes %>%
  write_csv(path="../output/MR_UN_graphs_node_annotation_2012.csv")

```

Edge Annotation
```{r}
blup.mr.graphs.UN.all %>% annotate.edges %>% 
  write_csv(path="../output/MR_UN_graphs_edge_annotation_all.csv")

blup.mr.graphs.UN.2011 %>% annotate.edges %>%
  write_csv(path="../output/MR_UN_graphs_edge_annotation_2011.csv")

blup.mr.graphs.UN.2012 %>% annotate.edges %>%
  write_csv(path="../output/MR_UN_graphs_edge_annotation_2012.csv")
```



## overlap between 2011 and 2012

Want to determine, for each FVT, how similar the connections are in 2011 and 2012

```{r, eval=FALSE}
MR35 <- blup.mr.graphs.UN[["blup.mr.graph.UN.35"]]$graph

sapply(blup.mr.graphs.UN, function(g) {
  
  sapply(FVTs, function(f) {
    V2011 <- V(g$graph)[.nei(str_c(f,"2011",sep = "."))] %>% names() %>% str_subset("Bra")
    V2012 <- V(g$graph)[.nei(str_c(f,"2012",sep = "."))] %>% names() %>% str_subset("Bra")
    length(intersect(V2011,V2012)) / length(union(V2011,V2012))
  })
})
```


## Permutations

What is a reasonable threshold?  To find out, permute blups relative to gene expression, make graphs at different thresholds, and return number of genes in network

```{r}
permute.graphs <- function(expr.blup,cutoffs=cutoffs,permute=TRUE) {
  blups <- expr.blup %>% select(Line, starts_with("UN"))
  expression <- expr.blup %>% select(Line, matches("^Bra[0-9]"))
  if(permute) blups$Line <- sample(blups$Line,replace = FALSE) # permute!
  expr.blup <- full_join(blups,expression, by = "Line")
  expr.blup.mr <- expr.blup[,-1] %>% 
    cor %>%  
    apply(2,function(x) rank(-abs(x))) %>%
    multiply_by(t(.)) %>%
    sqrt()
  blup.mr.graphs <- lapply(cutoffs, get.mr.subgraph, mr.matrix = expr.blup.mr, neighborhood = colnames(blups)[-1])
  sapply(blup.mr.graphs,function(graph) sum(grepl("^Bra",graph$nodes)))
}

system.time(test <- expr.blup.UN %>% 
  select(Line, ends_with("12"), starts_with("Bra"))  %>% 
  permute.graphs(cutoffs=c(10,20,30,50),permute=FALSE)) # 70 seconds

test
# should be the same as the first 4 of

sapply(blup.mr.graphs.UN.2012,function(graph) sum(grepl("^Bra",graph$nodes)))

```


```{r, eval=FALSE}
permutations <- 100

expr.blup.UN.12 <- expr.blup.UN %>% 
  select(Line, ends_with("12"), starts_with("Bra"))

#sfInit(parallel=TRUE,cpus = 4)

sfLibrary(igraph)
sfLibrary(magrittr)
sfLibrary(tidyverse)
sfExport("expr.blup.UN.12","permute.graphs","get.mr.subgraph","cutoffs")
system.time(perm.results.UN.100 <- sfSapply(1:permutations, function(p) {
  permute.graphs(expr.blup.UN.12,cutoffs=cutoffs)
}))

save(perm.results.UN.100, file="../output/permutation_results_UN.Rdata")

sfStop()
```

```{r}
load("../output/permutation_results_UN.Rdata")
apply(perm.results.UN.100,1,quantile,.95)
```

```{r}
kable(cbind(threshold=cutoffs,
            "0.95 expectation"=apply(perm.results.UN.100,1,quantile,.95),
            observed=sapply(blup.mr.graphs.UN.2012,function(graph) sum(grepl("^Bra",graph$nodes)))
), caption="Permutation analysis: UN network",digits=0)
```

Our networks are significant, but not sure that this is not helping set a threshold.  Try correlation with scale-free?

## Color edge by correlation direction
