==> scripts/05_WGCNA_UN.Rmd <==
---
title: "WGCNA UN"
output: html_notebook
---

WGCNA analysis of UN gene expression in B. rapa RILs.

Lets start with some notes on what WGCNA is doing.  (For more details, see [Zhang and Horvath, 2005](http://www.degruyter.com/view/j/sagmb.2005.4.1/sagmb.2005.4.1.1128/sagmb.2005.4.1.1128.xml;jsessionid=9F14E37B32BBAEB8DD0286D168168E7F) )

1. A correlation coefficient is calculated between all gene pairs.
2. This is converted to an adjacency matrix.  Unlike typical adjacency matrices (which have values of 0, meaning not connected, or 1, meaning connected, based on some hard threshold) here the adjacency matrix can take any value from 0 to 1.  Thus there is a soft threshold and the connectivity in the adjacency matrix is weighted by the correlations.  Zhang and Horvath discuss different ways to relate the correlation matrix to the adjacency matrix, the default in WGCNA, and what I use, is a power function where the adjacency represent the correlation coefficient raised to some power $\beta$.
3. The power, $\beta$, is chosen such that the (weighted) connectivity shows a power law topology i.e. R2 of log10(p(k)) ~ log10(k) is > 0.9
4. Once an adjacency matrix is calculated the the similarity of nodes in the matrix is calculated using a topological overlap matrix (TOM).  See Zhang and Horvath, 2005, for details.  But basically a two nodes have a high TOM if they are connected to each other and to a similar set of other nodes.
5. To define modules genes are hiearchicaly clustered based on their distance (1-TOM) and a dynamic method is used to define cutoffs from this tree and the dissimilarity matrix.  For more info see [here](https://labs.genetics.ucla.edu/horvath/htdocs/CoexpressionNetwork/BranchCutting/)
6.  I tshould be noted that Zhang and Horvath provide formula to generalize most network calcualtions (i.e. connectivity, etc) from a 0,1 adjacency matrix to a 0-1 adjacency matrix.

```{r}
library(tidyverse)
library(WGCNA)
library(ggplot2)
library(edgeR)
library(magrittr)
library(gplots)
options(stringsAsFactors = FALSE)
```

Get data
```{r}
#box_load(162362592785)
load("../output/Voom_UN.Rdata")
annotation  <- read.csv("../input/Brapa_V1.5_annotated.csv")
```

```{r}
RIL.coefs.UN <- grep("RIL",colnames(coef(voom.fit.UN)))
RIL.pvals.UN <- topTable(voom.fit.UN,coef = RIL.coefs.UN,number = Inf)
sum(RIL.pvals.UN$adj.P.Val<0.01)
sum(RIL.pvals.UN$adj.P.Val<10e-05) 
```
So lots of genes are highly significant

Start by taking the top 10000

```{r}
genes.of.interest.UN <- row.names(RIL.pvals.UN)[1:10000]
expr.data.UN <- voom.fit.UN$coefficients[genes.of.interest.UN,]
dim(expr.data.UN)
head(expr.data.UN[,1:6])
```


```{r}
expr.data.UN[,-1] <- expr.data.UN[,-1] + expr.data.UN[,1] # add the intercept to the coefficients
colnames(expr.data.UN) <- sub("pdata.UN$RIL","",colnames(expr.data.UN),fixed=TRUE)
colnames(expr.data.UN)[1] <- "RIL_1"
head(expr.data.UN[,1:6])
```

Get Rob's Data; remove STP Blup and other un-needed columns
```{r}
blups2011 <- read.csv("../input/Brapa2011BayesHeight_blups.csv",row.names = 1)
blups2012 <- read.csv("../input/Brapa2012BayesHeight_blups.csv", row.names = 1)
blups2011$Line %<>% paste("RIL",.,sep="_")
blups2012$Line %<>% paste("RIL",.,sep="_")
blups2011 %<>% subset(select=!grepl("V1|individual|blk|trt|treat|STP",colnames(.))) %>%
  rename(UNiD=UNInflect_DD,UNd=UNduration) %>%
  select(starts_with("Line"),starts_with("UN"),-matches("Inflect.*_size"))
blups2012 %<>% subset(select=!grepl("V1|individual|blk|trt|treat|STP",colnames(.))) %>%
  rename(UNiD=UNInflection_DD,UNd=UNduration) %>%
    select(starts_with("Line"),starts_with("UN"),-matches("Inflect.*_size"))
head(blups2011)
head(blups2012)
```

Have decided to use 2011 and 2012 BLUPs but UN only...
```{r}
blups.UN <- merge(blups2011,blups2012,by="Line",suffixes = c("11","12")) %>%
  select(starts_with("Line"),starts_with("UN"),-matches("Inflect.*_size"))
head(blups.UN)
```

## WGCNA 

### UN

transform
```{r}
head(expr.data.UN[,1:6])
expr.data.UN.t <- t(expr.data.UN)
head(expr.data.UN.t[,1:6])
```

check sample quality

```{r}
gag.UN <- goodSamplesGenes(expr.data.UN.t, verbose = 3)
gag.UN$allOK
```

cluster samples to look for outliers

```{r,fig.width=12}
sampleTREE.UN <- hclust(dist(expr.data.UN.t), method = "average")
plot(sampleTREE.UN,cex=.6)
heatmap.2(expr.data.UN.t,Rowv=as.dendrogram(sampleTREE.UN), scale="col", trace="none")
```
So a few samples are different but they don't seem crazy different; leave them in for now.

Soft thresholding
```{r}
powers <- c(c(1:10), seq(from = 12, to=20, by=2))
sft <- pickSoftThreshold(expr.data.UN.t, powerVector = powers, verbose = 5,networkType = "signed hybrid")
```

```{r}
sizeGrWindow(9, 5)
par(mfrow = c(1,2))
cex1 <- 0.9
# Scale-free topology fit index as a function of the soft-thresholding power
plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
     xlab="Soft Threshold (power)",ylab="Scale Free Topology Model Fit,signed R^2",type="n",
     main = paste("Scale independence"))
text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
     labels=powers,cex=cex1,col="red")
# this line corresponds to using an R^2 cut-off of h
abline(h=0.90,col="red")
# Mean connectivity as a function of the soft-thresholding power
plot(sft$fitIndices[,1], sft$fitIndices[,5],
     xlab="Soft Threshold (power)",ylab="Mean Connectivity", type="n",
     main = paste("Mean connectivity"))
text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=cex1,col="red")
```

choose 3

```{r}
softPower <- 3
adjacency <- adjacency(expr.data.UN.t, power = softPower, type = "signed hybrid")
# Turn adjacency into topological overlap
TOM <- TOMsimilarity(adjacency);
dissTOM <- 1-TOM
```

```{r}
# Call the hierarchical clustering function
geneTree <- hclust(as.dist(dissTOM), method = "average")
# Plot the resulting clustering tree (dendrogram)
sizeGrWindow(12,9)
plot(geneTree, xlab="", sub="", main = "Gene clustering on TOM-based dissimilarity",
     labels = FALSE, hang = 0.04)
```

define modules

```{r}
# We like large modules, so we set the minimum module size relatively high:
minModuleSize <- 30;
# Module identification using dynamic tree cut:
dynamicMods <- cutreeDynamic(dendro = geneTree, distM = dissTOM,
                             deepSplit <- 2, pamRespectsDendro = FALSE,
                             minClusterSize = minModuleSize);
table(dynamicMods)
```

```{r}
# Convert numeric labels into colors
dynamicColors = labels2colors(dynamicMods)
table(dynamicColors)
# Plot the dendrogram and colors underneath
sizeGrWindow(8,6)
plotDendroAndColors(geneTree, dynamicColors, "Dynamic Tree Cut",
                    dendroLabels = FALSE, hang = 0.03,
                    addGuide = TRUE, guideHang = 0.05,
                    main = "Gene dendrogram and module colors")
```

merge similar modules

```{r}
# Calculate eigengenes
MEList <- moduleEigengenes(expr.data.UN.t, colors = dynamicColors)
MEs <- MEList$eigengenes
# Calculate dissimilarity of module eigengenes
MEDiss <- 1-cor(MEs);
# Cluster module eigengenes
METree <- hclust(as.dist(MEDiss), method = "average");
# Plot the result
sizeGrWindow(7, 6)
plot(METree, main = "Clustering of module eigengenes",
     xlab = "", sub = "")
```

merge with correlation > 0.8
```{r}
MEDissThres = 0.2
# Plot the cut line into the dendrogram
plot(METree, main = "Clustering of module eigengenes",
     xlab = "", sub = "")
abline(h=MEDissThres, col = "red")
# Call an automatic merging function
merge = mergeCloseModules(expr.data.UN.t, dynamicColors, cutHeight = MEDissThres, verbose = 3)
# The merged module colors
mergedColors = merge$colors
# Eigengenes of the new merged modules:
mergedMEs = merge$newMEs
```

compare pre and post merge
```{r}
sizeGrWindow(12, 9)
#pdf(file = "Plots/geneDendro-3.pdf", wi = 9, he = 6)
plotDendroAndColors(geneTree, cbind(dynamicColors, mergedColors),
                    c("Dynamic Tree Cut", "Merged dynamic"),
                    dendroLabels = FALSE, hang = 0.03,
                    addGuide = TRUE, guideHang = 0.05)
#dev.off()
```

```{r}
# Rename to moduleColors
moduleColors = mergedColors
# Construct numerical labels corresponding to the colors
colorOrder = c("grey", standardColors(50));
moduleLabels = match(moduleColors, colorOrder)-1;
MEs = mergedMEs 
table(merge$colors)
length(table(merge$colors))
median(table(merge$colors))

```

```{r}
head(MEs)
dim(MEs)
dim(blups.UN)
rownames(expr.data.UN.t)
rownames(MEs)
rownames(MEs) <- rownames(expr.data.UN.t)
head(blups.UN)
```

```{r}
ME.blups <- merge(blups.UN,MEs,by.x=1.,by.y=0)
head(ME.blups)
```

```{r}
blups.cor <- cor(select(ME.blups,starts_with("UN")),
                 select(ME.blups, starts_with("ME")))
range(blups.cor)
blups.cor
```


```{r}
blups.cor.P <- corPvalueStudent(blups.cor,nrow(expr.data.UN.t)) %>%
  p.adjust() %>% #using "holm" method.  For relaxed stringency could use "fdr"
  matrix(nrow=nrow(blups.cor),
         dimnames = dimnames(blups.cor))

sum(blups.cor.P < 0.05)

blups.cor.sig <- blups.cor
blups.cor.sig[blups.cor.P>0.05] <- NA
blups.cor.sig <- blups.cor.sig[,!apply(blups.cor.sig,2,function(x) all(is.na(x)))]
blups.cor.sig

p.asterisk <- ifelse(blups.cor.P < 0.05, "*","")
```

plot it
```{r,fig.width=12}
# Will display correlations and their p-values
par(mar = c(8, 8, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = blups.cor,
               yLabels = names(blups.UN)[-1],
               xLabels = names(MEs) %>% str_replace("^ME",""),
               xSymbols = names(MEs),
               colorLabels = FALSE,
               colors = blueWhiteRed(50),
               textMatrix = p.asterisk,
               setStdMargins = FALSE,
               #cex.text = 0.5,
               zlim = c(-1,1),
               main = "UN Module-trait relationships")

pdf("../output/Module-trait_heatmap_UN_all.pdf",height = 8,width = 12)
par(mar = c(8, 8, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = blups.cor,
               yLabels = names(blups.UN)[-1],
               xLabels = names(MEs) %>% str_replace("^ME",""),
               xSymbols = names(MEs),
               colorLabels = FALSE,
               colors = blueWhiteRed(50),
               textMatrix = p.asterisk,
               setStdMargins = FALSE,
               #cex.text = 0.5,
               zlim = c(-1,1),
               main = paste("UN Module-trait relationships"))
dev.off()
```

Arbitrary, but let's take the max and min for each trait (so long as they are significant)
```{r}
blups.cor.5 <- blups.cor.sig

cor.top <- t(apply(blups.cor.5,1,function(x) {
  maxx = max(x,na.rm=TRUE)
  minx = min(x,na.rm=TRUE)
  ifelse(x == maxx | x == minx, x, NA)
}
))


cor.top <- cor.top[,apply(cor.top,2,function(x) !all(is.na(x)))]
cor.top
write.csv(cor.top,"../output/Eigen_trait_cor_UN_.5_threshold_all.csv")
```

write the significant Eigen genes
```{r}
head(MEs[,colnames(blups.cor.sig)])
write.csv(MEs[,colnames(blups.cor.sig)],"../output/Sig_Eigen_genes_UN_all.csv")
```

write all Eigen genes

```{r}
write.csv(MEs,"../output/All_Eigen_genes_UN_all.csv")
```

save WGCNA info

```{r}
save(annotation,blups.cor,blups.cor.5,blups.cor.P,blups.cor.sig,moduleColors,MEs,merge,expr.data.UN.t,
     file="../output/WGCNA_UN_all.Rdata")
```

## 2012 only

```{r}
#load("../output/WGCNA_UN_all.Rdata")
ME.blups <- blups.UN %>% select("Line", ends_with("12")) %>% merge(MEs,by.x=1.,by.y=0)
head(ME.blups)
```

```{r}
blups.cor <- cor(select(ME.blups,starts_with("UN")),
                 select(ME.blups, starts_with("ME")))
range(blups.cor)
blups.cor
```


```{r}
blups.cor.P <- corPvalueStudent(blups.cor,nrow(expr.data.UN.t)) %>%
  p.adjust() %>% #using "holm" method.  For relaxed stringency could use "fdr"
  matrix(nrow=nrow(blups.cor),
         dimnames = dimnames(blups.cor))

sum(blups.cor.P < 0.05)

blups.cor.sig <- blups.cor
blups.cor.sig[blups.cor.P>0.05] <- NA
blups.cor.sig <- blups.cor.sig[,!apply(blups.cor.sig,2,function(x) all(is.na(x)))]
blups.cor.sig

p.asterisk <- ifelse(blups.cor.P < 0.05, "*","")
```

plot it
```{r,fig.width=12}
# Will display correlations and their p-values
par(mar = c(8, 8, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = blups.cor,
               yLabels = rownames(blups.cor),
               xLabels = names(MEs),
               xSymbols = names(MEs) %>% str_replace("^ME",""),
               xColorLabels = TRUE,
               colors = blueWhiteRed(50),
               textMatrix = p.asterisk,
               setStdMargins = FALSE,
               #cex.text = 0.5,
               zlim = c(-1,1),
               main = "UN Module-trait relationships")

pdf("../output/Module-trait_heatmap_UN_2012.pdf",height = 4,width = 12)
par(mar = c(8, 8, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = blups.cor,
               yLabels = rownames(blups.cor),
               xLabels = names(MEs),
               xSymbols = names(MEs) %>% str_replace("^ME",""),
               XcolorLabels = FALSE,
               colors = blueWhiteRed(50),
               textMatrix = p.asterisk,
               setStdMargins = FALSE,
               #cex.text = 0.5,
               zlim = c(-1,1),
               main = paste("UN Module-trait relationships"))
dev.off()
```

Arbitrary, but let's take the max and min for each trait (so long as they are significant)
```{r}
blups.cor.5 <- blups.cor.sig

cor.top <- t(apply(blups.cor.5,1,function(x) {
  maxx = max(x,na.rm=TRUE)
  minx = min(x,na.rm=TRUE)
  ifelse(x == maxx | x == minx, x, NA)
}
))


cor.top <- cor.top[,apply(cor.top,2,function(x) !all(is.na(x)))]
cor.top
write.csv(cor.top,"../output/Eigen_trait_cor_UN_.5_threshold_2012.csv")
```

write the significant Eigen genes
```{r}
head(MEs[,colnames(blups.cor.sig)])
write.csv(MEs[,colnames(blups.cor.sig)],"../output/Sig_Eigen_genes_UN_2012.csv")
```

write all Eigen genes

```{r}
write.csv(MEs,"../output/All_Eigen_genes_UN_2012.csv")
```

save WGCNA info

```{r}
save(annotation,blups.cor,blups.cor.5,blups.cor.P,blups.cor.sig,moduleColors,MEs,merge,expr.data.UN.t,
     file="../output/WGCNA_UN_2012.Rdata")
```

## GO enrichment for each signficant cluster

```{r}
load("../output/WGCNA_UN_2012.Rdata")
module_genes <- tibble(GeneID=colnames(expr.data.UN.t), module=moduleColors)
module_genes
```

Get GO list and gene lengths
```{r, engine='bash', eval=FALSE}
cd ../input
wget http://www.g3journal.org/content/suppl/2014/08/12/g3.114.012526.DC1/FileS11.txt
wget http://jnmaloof.github.io/BIS180L_web/data/Brapa_CDS_lengths.txt
```


```{r}
library(goseq)
library(tidyverse)
library(stringr)
```

Format data for GOseq
```{r}
go.terms <- read.delim("../input/FileS11.txt",header=FALSE,as.is=TRUE)
head(go.terms)
names(go.terms) <- c("GeneID","GO")
summary(go.terms)

gene.lengths <- read_tsv("../input/Brapa_CDS_lengths.txt")

head(gene.lengths)
summary(gene.lengths)

#For this analysis the "Universe" will be all 10000 genes used as input to WGCNA, and the test set will be each module that showed a significant correlation with a FVT blup.

#we need to reduce the gene.length data to only contain entries for those genes in our universe.  We also need this as a vector
gene.lengths.module <- gene.lengths %>% 
  semi_join(module_genes,by="GeneID")

gene.lengths.vector <- as.vector(gene.lengths.module$Length)
names(gene.lengths.vector) <- gene.lengths.module$GeneID

#Do the reverse to make sure everything matches up (it seems that we don't have length info for some genes?)

module_genes <- semi_join(module_genes,gene.lengths.module)
```

Format go.terms for goseq. We want them in list format, and we need to separate the terms into separate elements.
```{r}
go.list <- strsplit(go.terms$GO,split=",")
names(go.list) <- go.terms$GeneID
head(go.list)
```

Now we will loop through each significant module

First, get the sig modules
```{r}
sig.modules <- colnames(blups.cor.P) %>%
  magrittr::extract(apply(blups.cor.P,2,function(x) any(x < 0.05))) %>%
  str_replace("ME","")
sig.modules
```


Format module data for goseq. We need a vector for each gene with 1 indicating module membership and 0 indicating not in module 
```{r}
file.remove("../output/SupTable_JM5_WGCNA_UN_GO_2012.csv")

GO.results <- lapply(sig.modules, function(module) {
  module01 <- module_genes$module %>% str_detect(module) %>% as.numeric()
  names(module01) <- module_genes$GeneID 
  
  #determines if there is bias due to gene length.
  nullp.result.tmp <- nullp(DEgenes = module01,bias.data = gene.lengths.vector,plot.fit = FALSE)
  
  #calculate p-values for each GO term
  GO.out.tmp <- goseq(pwf = nullp.result.tmp,gene2cat = go.list,test.cats=("GO:BP"))
  
  #Keep BP only
  GO.out.tmp <- GO.out.tmp[GO.out.tmp$ontology=="BP",]
  
  #Calculate FDR
  GO.out.tmp <- GO.out.tmp %>% as.tibble() %>%
    mutate(FDR=p.adjust(over_represented_pvalue, method = "fdr"),module=module) %>%
    filter(FDR < 0.05) %>%
    select(module,term,ontology,FDR,over_represented_pvalue,everything()) 
  
  write_csv(GO.out.tmp,path = "../output/SupTable_JM5_WGCNA_UN_GO_2012.csv",append = TRUE)
  
  GO.out.tmp
})
names(GO.results) <- sig.modules
```


```{r}
GO.results
```

## Are any of the clusters due to linkage?

```{r}
BrapaLocation <- read_csv("../input/Brapa_V1.5_annotated.csv") %>% select(GeneID=name, chrom, start) 
BrapaLocation
```


```{r}
modules_genes.pl <- left_join(module_genes, BrapaLocation) 
modules_genes.pl %>%
  filter(module %in% sig.modules, str_detect(chrom,"^A[01]")) %>%
  ggplot(aes(x=start)) +
  geom_dotplot() +
  facet_grid(module ~ chrom) +
  theme(axis.text.x = element_text(angle = 90),
        strip.text.y = element_text(angle=0))
ggsave("../output/WGCNA_2012_UN_module_by_chromosome.pdf")
```
